==> 01_R/0_paths.R <==

path_to_box <- "~/Library/CloudStorage/Box-Box/"
# path_to_box <- "C:/Users/Natalie Gradwohl/Box/"
# path_to_box <- c("C:/Users/palolili23/Box/")

path_to_raw_data <- paste0(path_to_box,
                           "Asian_Americans_dementia_data/analysis_data_tables/")

path_to_analytic_data <- paste0(path_to_box,
                                "Asian_Americans_dementia_data/aa_sex_gender_dementia/")
